An experimental feature of syncing with other runmpv instances over the internet
This should've been a lot of commits, rather, than one. But it would be too
cumbersome to split this chunk of work. Let it be a fine example of doing things
wrong.

Here is a list of what this commit actually does.

1. Code, which sends commands is moved from implementations of MpvInstance to
MpvEvents, as there is no difference between windows and linux implementations.

2. runmpv now can subscribe to mpv property changes, with MpvEvents.observe.
This is implemented via mpv observe_property.

3. runmpv connects to remote server if sync feature toggle is on. If mpv is
paused, then runmpv sends pause command to remote server, which is supposed
to broadcast that command to other mpv instances.

Also, runmpv listens to remote server and if pause command is given, runmpv
sends pause command to runmpv.

If runmpv couldn't connect to a remote server, it quits after video is started,
which is how it worked always, up until now. runmpv still quits after 10 seconds,
so you have to change that timeout manually to try sync mode.

Remote server address is hardcoded to localhost port 5454.

Code for remote server is not complete yet. It might be released as a separate
project, as there's no need to pack it as a static executable.
Or it's going to be a module after I split runmpv into several modules.

4. MpvCallback now accepts 3 parameters. Response body, MpvEvents and server
message queue.