Use automatically reconnecting byte channel in case there are disconnects
It's important to remember, that if you unplug network cord, then Channel.isOpen
still returns true. So the only way to know you're disconnected is to try writing
(reading will not do the trick) to socket and catching IOException.

What's even more interesting is that even if you unplug the cord,
WritableByteChannel.write still throws no exceptions until socket write buffer
is full.

But if other side is disconnected properly, then WritableByteChannel.write will
throw an exception, even if you didn't handle the disconnect.

ReadableByteChannel.read, on the other hand will not throw exceptions if you
unplug the cord and if server is disconnected. So, if you want to automatically
test if connection is active, before reading (to know, that reading really does
something), then you have to try writing before reading. And establish connection
again if write threw IOException. That will help in case of proper disconnect.
Not sure about just unplugging the cord though.