Only try focusing mpv if mpv is already working
If mpv is not open, then runmpv will start a new instance of mpv
and then mpv will take care of focusing it's own window.

runmpv only needs to focus mpv window if mpv already plays a video
and runmpv just sends a command to start playing a new one.