Do not try to call xdotool again and again if executable is not found
runmpv relies on xdotool to focus mpv window. Sometimes, runmpv tried to call
xdotool before mpv window was registered. And that caused xdotool exit with error.
So, I have decided to repeatedly call xdotool, until it was a success.

This all happened when mpv wasn't open yet and runmpv had to start it.
But, it's pointless to focus mpv window, when mpv is starting, because mpv
will do it itself.

So, runmpv was fixed not to try focusing mpv, if mpv is not already running.

But this feature of repeatedly calling xdotool left. If xdotool was not installed,
runmpv kept calling xdotool for 4 seconds, until it finally understood it's not
going to work. And only after that runmpv sent mpv a command to open a new video.

Because of that it took 4 seconds to start a new video with runmpv, if xdotool
wasn't installed.

If xdotool was not installed, then runmpv has been trying to
call it again and again to focus.

The solution is obviously to only call xdotool once and write logs, if executable
couldn't be found.