Fix runmpv hanging if exception occurred
runmpv is using a lock, which is released only after config is read and
serverChannel is created. Main thread waits until the lock is released.

But an exception could happen, before lock is released and in that case you
have to release the lock inside catch or finally block.

That is actually what this commit fixes.