Take path to logging.properties from logSettings property
If path in logSettings is relative, then runmpv resolves it relative
to <executableDir>.

logSettings includes file name, so it is not necessary to name the file
"logging.properties". logSettings is taken from command line arguments
first. If it's not there, then logging.properties is taken from default
location. Default location is <executableDir>/logging.properties . This
file is included in binary distribution.

If, after reading full config runmpv understands, that logSettings is also
defined if runmpv.properties, it restarts logging system, using settings
from runmpv.properties. But it only occurs if logSettings is absent in
command line arguments.